Changing search bars into dropdown bars, very early version
