+++ app/search/styles.css
@@ -2,100 +2,24 @@ h2 {
     text-align: center;
 }
 
-.filter-container {
+.arrow-down::before {
+    content: "⌄";
+}
+
+.arrow-up::before {
+    content: "⌃";
+}
+
+.dropdown-container .dropdown {
     position: relative;
 }
 
-.filter-options {
-    display: flex;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.filter-options button {
-    margin-left: 15px;
-    margin-right: 15px;
-    width: 120px;
-    position: relative;
-}
-
-.filter-options b {
-    margin-right: 5px;
-}
-
-.filter-options button::after {
-    content: "";
-    position: absolute;
-    top: 50%;
-    right: 3px;
-    width: 0;
-    height: 0;
-    border: 4px solid transparent;
-    border-top-color: black;
-    transform: translateY(-50%);
-    transition: transform 0.3s ease;
-}
-
-.filter-options button.open::after {
-    transform: translateY(-50%) rotate(180deg);
-}
-
-.filter-button-wrapper {
-    position: relative;
-}
-
-.options-dropdown {
+.dropdown-container .dropdown-menu {
     position: absolute;
     top: 100%;
     left: 0;
-    background: white;
-    border: 1px solid #ccc;
-    padding: 10px;
-    z-index: 10;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-    width: 100px;
-    margin-left: 15px;
+    right: 0;
+    margin: 0;
+    border-radius: 0;
 }
 
-.search-bar-container {
-    display: flex;
-    flex-direction: column;
-    width: 100%;
-    max-width: 600px;
-    margin: 0 auto;
-}
-
-.search-bar {
-    display: flex;
-    align-items: center;
-    margin-bottom: 15px;
-}
-
-.search-input {
-    flex-grow: 1;
-    padding: 10px;
-    font-size: 16px;
-    border: 1px solid #ccc;
-    border-radius: 4px;
-}
-
-.search-input:first-of-type {
-    margin-right: 10px;
-}
-
-.search-button {
-    padding: 10px 20px;
-    font-size: 16px;
-    background-color: #3498db;
-    color: white;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-    align-self: flex-start; /* Align the button to the start of the flex container */
-    margin-bottom: 15px;
-}
-
-.search-button:hover {
-    background-color: #2980b9;
-}
